Add files via upload
Earthquake is a good description of a natural phenomenon which suddenly strikes an area causing damage that varies according to the intensity of the quake and local geological conditions.When a damaging earthquake hits a populated area, it is very important to grasp the overall distribution of building damage within several hours to a few days of the event, for the purpose of emergency response and restoration planning. An automated model that can predict the extent of damage that is done to a building post earthquake can readily reduce the manpower required in analysing such tasks and lead to quick decision making.